review: drop redundant half(0.0h) cast; fix stale invariant comment
- `half(0.0h)` in kv_tile tail-zero writes → `0.0h`. The `0.0h` literal
  is already half-precision; the explicit cast was noise.
- Rust dispatch invariant comment mentioned "max ~64 lanes wide" which
  was the Iter 2 scalar V-accum constraint (po[8] × 32 lanes = 256).
  Iter 4's MMA variant has no such limit; the real upper bound on hd
  is from the final normalize + write-out loops.

No behavior change. Parity test passes.

+++ wick/src/backend/shaders/attention_prefill.metal
@@ -146,7 +146,7 @@ kernel void attention_prefill(
             kv_tile[t * hd + d] = k_cache[(c0 + t) * kv_dim + kv_h_off + d];
         }
         for (uint idx = tid + c_len * hd; idx < C * hd; idx += N_THREADS) {
-            kv_tile[idx] = half(0.0h);
+            kv_tile[idx] = 0.0h;
         }
         threadgroup_barrier(mem_flags::mem_threadgroup);
 
@@ -208,7 +208,7 @@ kernel void attention_prefill(
             kv_tile[t * hd + d] = v_cache[(c0 + t) * kv_dim + kv_h_off + d];
         }
         for (uint idx = tid + c_len * hd; idx < C * hd; idx += N_THREADS) {
-            kv_tile[idx] = half(0.0h);
+            kv_tile[idx] = 0.0h;
         }
         threadgroup_barrier(mem_flags::mem_threadgroup);
